Split the sizing machinery into laws, context and field modules
hisim/config/sizing.py had grown to 641 lines carrying three separable
concerns. It now honors the ~500-line module rule as three modules in
dependency order: laws.py holds the SizingLaw expression algebra and the
sizing errors (how a value derives), context.py holds SizingContext and the
Size term vocabulary side by side (what laws read - keeping the single
registry reviewable in one file), and sizing.py keeps the field machinery
(AUTO, sized_field, resolve_config - how fields declare and resolve).
The public API through hisim.config is unchanged; base.py, engine.py and the
sizing tests now import each name from its actual home, per the no-alias
convention of the package.
